refactor popup.mjs into submodules
