+++ src/popup/address-book-list.mjs
@@ -6,7 +6,11 @@ function writeAddressBookElement(addressBook, index) {
   return `<li data-address-book="${addressBook.id}" ${className}>${name}</li>`;
 }
 
-export function createAddressBookList({ data, click }) {
+export function createAddressBookList({
+  data,
+  state,
+  components: { categoryTitle, contactList, categoryTree },
+}) {
   let component = new Component({
     element: "#address-books",
     data,
@@ -15,12 +19,27 @@ export function createAddressBookList({ data, click }) {
       return elements;
     },
   });
-  click && component.element.addEventListener("click", click);
-  component.element.addEventListener("click", ({ target }) => {
+  async function click({ target }) {
+    const addressBookId = target.dataset.addressBook;
+    if (addressBookId == null) return;
+    state.currentAddressBook = state.addressBooks.get(addressBookId);
+    state.currentCategoryElement = null;
+    categoryTitle.innerText = state.currentAddressBook.name;
     for (const e of target.parentElement.children) {
       e.classList.remove("selected");
     }
     target.classList.toggle("selected");
-  });
+    return Promise.all([
+      categoryTree.update({
+        addressBook: state.currentAddressBook,
+        activeCategory: null,
+      }),
+      contactList.update({
+        addressBook: state.currentAddressBook,
+        contacts: state.currentAddressBook.contacts,
+      }),
+    ]);
+  }
+  component.element.addEventListener("click", click);
   return component;
 }
